+++ packages/expo-location/ios/LocationModule.swift
@@ -281,5 +281,31 @@ public final class LocationModule: Module {
     AsyncFunction("hasStartedGeofencingAsync") { (taskName: String) -> Bool in
       return try taskManager.task(withName: taskName, hasConsumerOf: EXGeofencingTaskConsumer.self)
     }
+
+    // Background motion activity
+
+    AsyncFunction("startMotionActivityUpdatesAsync") { (taskName: String) in
+      guard CMMotionActivityManager.isActivityAvailable() else {
+        throw Exceptions.MotionActivityUnavailable()
+      }
+      let authorizationStatus = CMMotionActivityManager.authorizationStatus()
+      guard authorizationStatus != .denied && authorizationStatus != .restricted else {
+        throw Exceptions.MotionActivityUnauthorized()
+      }
+
+      try taskManager.registerTask(withName: taskName, consumer: MotionActivityTaskConsumer.self, options: [:])
+    }
+
+    AsyncFunction("stopMotionActivityUpdatesAsync") { (taskName: String) in
+      let taskManager = try taskManager
+
+      try EXUtilities.catchException {
+        taskManager.unregisterTask(withName: taskName, consumerClass: MotionActivityTaskConsumer.self)
+      }
+    }
+
+    AsyncFunction("hasStartedMotionActivityUpdatesAsync") { (taskName: String) -> Bool in
+      return try taskManager.task(withName: taskName, hasConsumerOf: MotionActivityTaskConsumer.self)
+    }
   }
 }